// src/generated/serverApi/v3/models/class-info-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Schulcloud-Verbund-Software Server API
 * This is v3 of Schulcloud-Verbund-Software Server. Checkout /docs for v1.
 *
 * The version of the OpenAPI document: 3.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { CourseInfoResponse } from './course-info-response';

/**
 * 
 * @export
 * @interface ClassInfoResponse
 */
export interface ClassInfoResponse {
    /**
     * 
     * @type {string}
     * @memberof ClassInfoResponse
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof ClassInfoResponse
     */
    type: ClassInfoResponseType;
    /**
     * 
     * @type {string}
     * @memberof ClassInfoResponse
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof ClassInfoResponse
     */
    externalSourceName?: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof ClassInfoResponse
     */
    teacherNames: Array<string>;
    /**
     * 
     * @type {string}
     * @memberof ClassInfoResponse
     */
    schoolYear?: string;
    /**
     * 
     * @type {boolean}
     * @memberof ClassInfoResponse
     */
    isUpgradable?: boolean;
    /**
     * 
     * @type {number}
     * @memberof ClassInfoResponse
     */
    studentCount: number;
    /**
     * 
     * @type {Array<CourseInfoResponse>}
     * @memberof ClassInfoResponse
     */
    synchronizedCourses?: Array<CourseInfoResponse>;
}

/**
    * @export
    * @enum {string}
    */
export enum ClassInfoResponseType {
    CLASS = 'class',
    GROUP = 'group'
}