// src/generated/serverApi/v3/models/course-info-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Schulcloud-Verbund-Software Server API
 * This is v3 of Schulcloud-Verbund-Software Server. Checkout /docs for v1.
 *
 * The version of the OpenAPI document: 3.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface CourseInfoResponse
 */
export interface CourseInfoResponse {
    /**
     * 
     * @type {string}
     * @memberof CourseInfoResponse
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof CourseInfoResponse
     */
    name: string;
}